+++ apps/vscode-extension/src/utils/getActionTemplate.ts
@@ -257,6 +257,12 @@ export const getActionTemplate = (selectedAction: QuickPickItem): any => {
     };
   }
 
+  if (action === Action.WaitForNext) {
+    return {
+      action: Action.WaitForNext,
+    };
+  }
+
   /**
    * VSCode actions
    */
@@ -306,52 +312,52 @@ export const getActionTemplate = (selectedAction: QuickPickItem): any => {
   /**
    * External applications
    */
-  if (action === Action.openPowerPoint) {
+  if (action === Action.OpenPowerPoint) {
     return {
-      action: Action.openPowerPoint,
+      action: Action.OpenPowerPoint,
     };
   }
 
-  if (action === Action.openKeynote) {
+  if (action === Action.OpenKeynote) {
     return {
-      action: Action.openKeynote,
+      action: Action.OpenKeynote,
     };
   }
 
   /**
    * GitHub Copilot actions
    */
-  if (action === Action.openChat) {
+  if (action === Action.OpenChat) {
     return {
-      action: Action.openChat,
+      action: Action.OpenChat,
     };
   }
-  if (action === Action.newChat) {
+  if (action === Action.NewChat) {
     return {
-      action: Action.newChat,
+      action: Action.NewChat,
     };
   }
-  if (action === Action.askChat) {
+  if (action === Action.AskChat) {
     return {
-      action: Action.askChat,
+      action: Action.AskChat,
       message: '',
     };
   }
-  if (action === Action.editChat) {
+  if (action === Action.EditChat) {
     return {
-      action: Action.editChat,
+      action: Action.EditChat,
       message: '',
     };
   }
-  if (action === Action.agentChat) {
+  if (action === Action.AgentChat) {
     return {
-      action: Action.agentChat,
+      action: Action.AgentChat,
       message: '',
     };
   }
-  if (action === Action.closeChat) {
+  if (action === Action.CloseChat) {
     return {
-      action: Action.closeChat,
+      action: Action.CloseChat,
     };
   }