+++ packages/common/src/models/Action.ts
@@ -32,6 +32,7 @@ export enum Action {
   // Timeout
   WaitForTimeout = 'waitForTimeout',
   WaitForInput = 'waitForInput',
+  WaitForNext = 'waitForNext',
   // VSCode
   ExecuteVSCodeCommand = 'executeVSCodeCommand',
   ShowInfoMessage = 'showInfoMessage',
@@ -45,15 +46,15 @@ export enum Action {
   // Extensibility
   Snippet = 'snippet',
   // External applications
-  openPowerPoint = 'openPowerPoint',
-  openKeynote = 'openKeynote',
+  OpenPowerPoint = 'openPowerPoint',
+  OpenKeynote = 'openKeynote',
   // GitHub Copilot
-  openChat = 'openCopilotChat',
-  newChat = 'newCopilotChat',
-  askChat = 'askCopilotChat',
-  editChat = 'editCopilotChat',
-  agentChat = 'agentCopilotChat',
-  closeChat = 'closeCopilotChat',
+  OpenChat = 'openCopilotChat',
+  NewChat = 'newCopilotChat',
+  AskChat = 'askCopilotChat',
+  EditChat = 'editCopilotChat',
+  AgentChat = 'agentCopilotChat',
+  CloseChat = 'closeCopilotChat',
   // Interaction
   TypeText = 'typeText',
   CopyToClipboard = 'copyToClipboard',